requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib==1.7.4
python-multipart==0.0.9
pytest==8.0.0
httpx==0.27.0